+++ packages/react-static-plugin-less/src/node.api.js
@@ -29,12 +29,6 @@ export default ({ includePaths = [], ...rest }) => ({
         plugins: () => [
           postcssFlexbugsFixes,
           autoprefixer({
-            browsers: [
-              '>1%',
-              'last 4 versions',
-              'Firefox ESR',
-              'not ie < 9', // React doesn't support IE8 anyway
-            ],
             flexbox: 'no-2009',
           }),
         ],