+++ packages/react-static-plugin-sass/src/node.api.js
@@ -28,12 +28,6 @@ export default ({ includePaths = [], ...rest }) => ({
         plugins: () => [
           postcssFlexbugsFixes,
           autoprefixer({
-            browsers: [
-              '>1%',
-              'last 4 versions',
-              'Firefox ESR',
-              'not ie < 9', // React doesn't support IE8 anyway
-            ],
             flexbox: 'no-2009',
           }),
         ],
@@ -75,7 +69,11 @@ export default ({ includePaths = [], ...rest }) => ({
       use: loaders,
     })
 
-    if (config.optimization.splitChunks.cacheGroups.styles) {
+    if (
+      config.optimization.splitChunks &&
+      config.optimization.splitChunks &&
+      config.optimization.splitChunks.cacheGroups.styles
+    ) {
       config.optimization.splitChunks.cacheGroups.styles.test = /\.(c|sc|sa)ss$/
     }