+++ packages/react-static/src/browser/hooks/usePrefetch.js
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react'
 import { getRoutePath } from '../utils'
-import { prefetch } from ".."
+import { prefetch } from '..'
 import onVisible from '../utils/Visibility'
 
 const usePrefetch = (path, ref = useRef()) => {