+++ packages/react-static/src/static/__mocks__/mock-plugin/node.api.js
@@ -1,0 +1,5 @@
+export default () => ({
+  webpack: config => {
+    return config
+  },
+})