+++ packages/react-static/src/static/getConfig.js
@@ -213,6 +213,8 @@ export function buildConfig(state, config = {}) {
   process.env.REACT_STATIC_DISABLE_RUNTIME = config.disableRuntime
   process.env.REACT_STATIC_PRELOAD_POLL_INTERVAL = config.preloadPollIntervalw
 
+  process.env.REACT_STATIC_ROOT_PATH_READ_ONLY = paths.ROOT
+
   process.env.REACT_STATIC_TEMPLATES_PATH = nodePath.join(
     paths.ARTIFACTS,
     'react-static-templates.js'
@@ -283,7 +285,7 @@ export function buildConfig(state, config = {}) {
       () => {
         if (process.env.NODE_ENV === 'test') {
           // Allow plugins to be mocked
-          return 'mock-plugin'
+          return require('path').resolve('./src/static/__mocks__/mock-plugin')
         }
       },
     ].reduce((prev, curr) => prev || curr(), null)
@@ -314,7 +316,9 @@ export function buildConfig(state, config = {}) {
         buildPluginHooks = require(nodeLocation).default
       } else if (originalLocation !== paths.ROOT && !browserLocation) {
         throw new Error(
-          `Could not find a valid node.api.js or browser.api.js plugin file in "${location}"`
+          `Could not find a valid node.api.js or browser.api.js plugin file in "${location}". \n` +
+            `The original location: "${originalLocation}". \n` +
+            `The root location: "${paths.ROOT}".`
         )
       }